Factor out fromrequest tests into separate module
